travis: Publishes latest tag only on stable releases

+++ .travis/build-images-frontend.sh
@@ -45,26 +45,27 @@ docker manifest annotate fragaria/karmen-frontend:$TRAVIS_BRANCH fragaria/karmen
 
 docker manifest push fragaria/karmen-frontend:$TRAVIS_BRANCH
 
-# latest
+# latest only if not rc
+if [[ "$TRAVIS_TAG" =~ ^v[0-9.]*$ ]]; then
+  docker pull fragaria/karmen-frontend:$TRAVIS_BRANCH
+  docker pull fragaria/karmen-frontend:$TRAVIS_BRANCH-amd64
+  docker pull fragaria/karmen-frontend:$TRAVIS_BRANCH-armhf
+  docker tag fragaria/karmen-frontend:$TRAVIS_BRANCH fragaria/karmen-frontend:latest
+  docker tag fragaria/karmen-frontend:$TRAVIS_BRANCH-amd64 fragaria/karmen-frontend:latest-amd64
+  docker tag fragaria/karmen-frontend:$TRAVIS_BRANCH-armhf fragaria/karmen-frontend:latest-armhf
+  docker push fragaria/karmen-frontend:latest
+  docker push fragaria/karmen-frontend:latest-armhf
+  docker push fragaria/karmen-frontend:latest-amd64
 
-docker pull fragaria/karmen-frontend:$TRAVIS_BRANCH
-docker pull fragaria/karmen-frontend:$TRAVIS_BRANCH-amd64
-docker pull fragaria/karmen-frontend:$TRAVIS_BRANCH-armhf
-docker tag fragaria/karmen-frontend:$TRAVIS_BRANCH fragaria/karmen-frontend:latest
-docker tag fragaria/karmen-frontend:$TRAVIS_BRANCH-amd64 fragaria/karmen-frontend:latest-amd64
-docker tag fragaria/karmen-frontend:$TRAVIS_BRANCH-armhf fragaria/karmen-frontend:latest-armhf
-docker push fragaria/karmen-frontend:latest
-docker push fragaria/karmen-frontend:latest-armhf
-docker push fragaria/karmen-frontend:latest-amd64
+  docker manifest create fragaria/karmen-frontend:latest \
+              fragaria/karmen-frontend:latest-amd64 \
+              fragaria/karmen-frontend:latest-armhf
 
-docker manifest create fragaria/karmen-frontend:latest \
-            fragaria/karmen-frontend:latest-amd64 \
-            fragaria/karmen-frontend:latest-armhf
+  docker manifest annotate fragaria/karmen-frontend:latest fragaria/karmen-frontend:latest-armhf --arch arm
+  docker manifest annotate fragaria/karmen-frontend:latest fragaria/karmen-frontend:latest-amd64 --arch amd64
 
-docker manifest annotate fragaria/karmen-frontend:latest fragaria/karmen-frontend:latest-armhf --arch arm
-docker manifest annotate fragaria/karmen-frontend:latest fragaria/karmen-frontend:latest-amd64 --arch amd64
-
-docker manifest push fragaria/karmen-frontend:latest
+  docker manifest push fragaria/karmen-frontend:latest
+fi
 
 # Delete unnecessary tags
 DOCKER_TOKEN=$(curl -s -H "Content-Type: application/json" -X POST -d '{"username": "'${DOCKER_USER}'", "password": "'${DOCKER_PASSWORD}'"}' https://hub.docker.com/v2/users/login/ | jq -r .token)
@@ -79,12 +80,14 @@ curl -i -X DELETE \
   -H "Authorization: JWT ${DOCKER_TOKEN}" \
   "https://hub.docker.com/v2/repositories/fragaria/karmen-frontend/tags/${TRAVIS_BRANCH}-armhf/"
 
-curl -i -X DELETE \
-  -H "Accept: application/json" \
-  -H "Authorization: JWT ${DOCKER_TOKEN}" \
-  "https://hub.docker.com/v2/repositories/fragaria/karmen-frontend/tags/latest-amd64/"
+if [[ "$TRAVIS_TAG" =~ ^v[0-9.]*$ ]]; then
+  curl -i -X DELETE \
+    -H "Accept: application/json" \
+    -H "Authorization: JWT ${DOCKER_TOKEN}" \
+    "https://hub.docker.com/v2/repositories/fragaria/karmen-frontend/tags/latest-amd64/"
 
-curl -i -X DELETE \
-  -H "Accept: application/json" \
-  -H "Authorization: JWT ${DOCKER_TOKEN}" \
-  "https://hub.docker.com/v2/repositories/fragaria/karmen-frontend/tags/latest-armhf/"
+  curl -i -X DELETE \
+    -H "Accept: application/json" \
+    -H "Authorization: JWT ${DOCKER_TOKEN}" \
+    "https://hub.docker.com/v2/repositories/fragaria/karmen-frontend/tags/latest-armhf/"
+fi
